+++ batchservice/src/main/java/nu/yona/server/batch/jobs/SendSystemMessageBatchJob.java
@@ -34,6 +34,7 @@ import org.springframework.stereotype.Component;
 
 import nu.yona.server.exceptions.YonaException;
 import nu.yona.server.messaging.entities.MessageDestination;
+import nu.yona.server.messaging.entities.MessageRepository;
 import nu.yona.server.messaging.entities.SystemMessage;
 import nu.yona.server.messaging.service.MessageService;
 import nu.yona.server.rest.RestUtil;
@@ -73,6 +74,9 @@ public class SendSystemMessageBatchJob
 	@Autowired
 	private MessageService messageService;
 
+	@Autowired
+	private MessageRepository messageRepository;
+
 	@Bean("sendSystemMessageJob")
 	public Job sendSystemMessagesBatchJob()
 	{
@@ -111,6 +115,7 @@ public class SendSystemMessageBatchJob
 						.getMessageDestination(userAnonymized.getAnonymousDestination().getId());
 				SystemMessage message = SystemMessage.createInstance(messageText);
 				messageDestination.send(message);
+				messageRepository.save(message);
 				messageService.sendFirebaseNotification(message, userAnonymized);
 
 				return messageDestination;
